requirements.txt
Django==4.0.4
django-heroku
gunicorn
asgiref
psycopg2
whitenoise

beautifulsoup4-4.11.1
lxml-4.9.1
requests-2.28.1


django-celery-beat
celery-5.2.7